Third push
Contain an example Excel file that contains all the necessary
information for each applicant. By running "mvn clean install"
which will produce the artifact, college-admission.jar in the target
folder and execute the jar by running "java -jar college-admission.jar"
will prompt you to provide an excel file.
For example, the user could type in the example file by typing:
"src/test/resources/record.xlsx", and the file will be written with
the result and the user should be able the see the result after opening
the file.